Makefile: ignore web preview/modules on modecheck
One Makefile target checks that there is only files that aren't
executable in the known hierarchy for the repo.

In case someone builds and serves the docs website however, some
directories get populated with JS scripts, which are also executable,
making it impossible to run the tests, as the executable files lifted
from the file hierarchy gets too large, and the test fails because of
too many arguments.

Since those files are compilation artifacts, they are safe to ignore, so
we ignore all the executable files below the website/node-modules and
website/website-preview directories.